rtl8188eu: Fix 4th argument to EFUSE_GetEfuseDefinition()
If this is other than u16, a BE machine gets the wrong answer.

+++ hal/hal_com.c
@@ -9234,7 +9234,7 @@ exit:
 void rtw_dump_cur_efuse(PADAPTER padapter)
 {
 	int i =0;
-	int mapsize =0;
+	u16 mapsize =0;
 	HAL_DATA_TYPE *hal_data = GET_HAL_DATA(padapter);
 
 	EFUSE_GetEfuseDefinition(padapter, EFUSE_WIFI, TYPE_EFUSE_MAP_LEN , (void *)&mapsize, _FALSE);
@@ -9255,7 +9255,7 @@ u32 Hal_readPGDataFromConfigFile(PADAPTER padapter)
 {
 	HAL_DATA_TYPE *hal_data = GET_HAL_DATA(padapter);
 	u32 ret = _FALSE;
-	u32 maplen = 0;
+	u16 maplen = 0;
 
 	EFUSE_GetEfuseDefinition(padapter, EFUSE_WIFI, TYPE_EFUSE_MAP_LEN , (void *)&maplen, _FALSE);
